Fix prettier for auto EOL, make universal menu routes interface, install react-hook-form
