+++ package.json
@@ -20,6 +20,7 @@
     "immer": "^10.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.53.0",
     "react-icons": "^4.9.0",
     "three": "0.153.0",
     "ts-pattern": "^4.3.0",
