+++ src/pages/GetSecretPage.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const GetSecretPage: React.FC = () => (
+  <div>Here will be secret page getting...</div>
+);
