+++ src/pages/index.ts
@@ -1,0 +1,2 @@
+export { GeneratePage } from './GeneratePage';
+export { GetSecretPage } from './GetSecretPage';
