+++ src/routes/index.ts
@@ -1,0 +1,2 @@
+export * from './AppRoutes';
+export * from './Menu';
